Move class 'ResponseGenerator' to a new file
