+++ src/infer_response.h
@@ -126,21 +126,4 @@ class InferResponse {
   bi::managed_external_buffer::handle_t next_response_handle_;
 };
 
-#ifdef TRITON_PB_STUB
-class ResponseGenerator {
- public:
-  ResponseGenerator(
-      const std::vector<std::shared_ptr<InferResponse>>& responses);
-
-  std::shared_ptr<InferResponse> Next();
-  int Length();
-  std::vector<std::shared_ptr<InferResponse>>::iterator Begin();
-  std::vector<std::shared_ptr<InferResponse>>::iterator End();
-
- private:
-  std::vector<std::shared_ptr<InferResponse>> responses_;
-  size_t index_;
-};
-#endif
-
 }}}  // namespace triton::backend::python
