+++ src/pb_generator.cc
@@ -1,0 +1,68 @@
+// Copyright 2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions
+// are met:
+//  * Redistributions of source code must retain the above copyright
+//    notice, this list of conditions and the following disclaimer.
+//  * Redistributions in binary form must reproduce the above copyright
+//    notice, this list of conditions and the following disclaimer in the
+//    documentation and/or other materials provided with the distribution.
+//  * Neither the name of NVIDIA CORPORATION nor the names of its
+//    contributors may be used to endorse or promote products derived
+//    from this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ``AS IS'' AND ANY
+// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+// PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+// OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+#include "pb_generator.h"
+
+#include <pybind11/embed.h>
+namespace py = pybind11;
+
+namespace triton { namespace backend { namespace python {
+
+ResponseGenerator::ResponseGenerator(
+    const std::vector<std::shared_ptr<InferResponse>>& responses)
+    : responses_(responses), index_(0)
+{
+}
+
+std::shared_ptr<InferResponse>
+ResponseGenerator::Next()
+{
+  if (index_ == responses_.size()) {
+    throw py::stop_iteration("Iteration is done for the responses.");
+  }
+
+  return responses_[index_++];
+}
+
+int
+ResponseGenerator::Length()
+{
+  return responses_.size();
+}
+
+std::vector<std::shared_ptr<InferResponse>>::iterator
+ResponseGenerator::Begin()
+{
+  return responses_.begin();
+}
+
+std::vector<std::shared_ptr<InferResponse>>::iterator
+ResponseGenerator::End()
+{
+  return responses_.end();
+}
+
+}}}  // namespace triton::backend::python
